==> notebooks/5.0_explanability.ipynb <==
import joblib
import pandas as pd
import xgboost
import shap

LABEL_COLUMN = "churned"

shap.initjs() 

# Load dataset

data = pd.read_csv('../data/users_train.csv')
x_test = data.drop(columns=LABEL_COLUMN)
y_test = data[LABEL_COLUMN]

# Load best model

pipeline = joblib.load('models/xgb_model_full.joblib')

preprocesor = pipeline['preprocessor']
model = pipeline['classifier']

# Prepare dataset

observations = preprocesor.transform(x_test)

column_name = (
    preprocesor.transformers_[0][2]
    + list(preprocesor.transformers_[1][1]['one-hot'].get_feature_names(preprocesor.transformers[1][2]))
)

observations_df = pd.DataFrame(observations, columns=column_name)

observations_df

# Interpretability

## Feature importance desde el modelo 

(
    pd.Series(model.feature_importances_, index=observations_df.columns)
    .sort_values(ascending=False)
    .iloc[:10]
    .plot(kind="bar")
)

## Shap Values

### TreeExplainer

explainer = shap.TreeExplainer(model)
shap_values = explainer.shap_values(observations)

shap.summary_plot(shap_values, observations_df, plot_type="bar")

### Explainer

explainer = shap.Explainer(model)
shap_values = explainer(observations_df)

shap_values

#### visualize the first prediction's explanation


shap.plots.waterfall(shap_values[0])

#### visualize all predictiosn

shap.plots.beeswarm(shap_values)

shap.plots.force(shap_values[0])

#shap.plots.force(explainer.expected_value, shap_values.values)
